Add single steps back into the route scoring in a limited way.
Previously, the single steps would be ignored entirely. This change
makes it so those odd steps are at least still weighted to be as late as
possible.

+++ src/engine.cc
@@ -233,7 +233,7 @@ int Engine::get_variable_count() const
 	return _parameters.randomizer->get_set_variable_count();
 }
 
-int Engine::get_score() const
+double Engine::get_score() const
 {
 	return _parameters.randomizer->get_score();
 }
